Refactor folder structure of user section pages
