+++ apps/developer-portal/src/pages/users/user-edit.tsx
@@ -21,11 +21,11 @@ import { UserAvatar } from "@wso2is/react-components";
 import React, { ReactElement, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useDispatch } from "react-redux";
-import { getUserDetails } from "../api";
-import { EditUser } from "../components";
-import { history } from "../helpers";
-import { PageLayout } from "../layouts";
-import { AlertInterface, BasicProfileInterface, createEmptyProfile } from "../models";
+import { getUserDetails } from "../../api";
+import { EditUser } from "../../components";
+import { history } from "../../helpers";
+import { PageLayout } from "../../layouts";
+import { AlertInterface, BasicProfileInterface, createEmptyProfile } from "../../models";
 
 /**
  * User Edit page.
